- convert_second_to_ms - always return int.

+++ core/utils/time_utils.py
@@ -1,7 +1,7 @@
 from time import time
 
 def convert_second_to_ms(time_in_second: int) -> int:
-    return time_in_second * 1000
+    return int(time_in_second * 1000)
 
 def convert_second_to_absolute_expiray_in_ms(time_in_second: int) -> int:
     return convert_second_to_ms(time()) + convert_second_to_ms(time_in_second)
